[perftest][cmake] Standardize on using ${source} instead of ${srcdir}/${module_name_path}.swift.
This is an intermediate commit in a series of commits towards being able to
build the perf test suite from *.sib files.

Doing so will enable us to:

1. Stress the serialization pipeline.
2. Provide an interesting test case for verifying that executables compiled
from *.sib yield the same object files as if we compiled directly to *.o files.

Keep in mind this is a longer term effort that I am doing on the side.