Modify computeSubSeqRelation to return true when both paths are empty.
Note: this is not the same as one path does not have value. A path can have no
value when we can not find a projection path between Base and Derived.

If we've found paths for both projections, and they turn'ed out to be both empty.
Its safe to say that they are accessing the same fields in the object.

I failed to create a test case as the changed code is currently not reached with
the current way we use it.

But it will be used as i use projection to do more analysis on epilogue release
matcher.