Create a fast path for release silargument.
When we have a single release that can be traced back to a SILArgument.
i.e. the released value is RC-identical to the SILArgument. we do not
need projection to check whether there are overlapping/uncovered releases
for the SILargument (which will result we exit the epilogue walking sequnece).

This brings the # of Total owned args -> guaranteed args from 118 and 149.

There are 23 owned args which we can not find epilogue releases for at this
point and many (if not most) is a result of partial_apply which rc-identity
nor projection can handle.

with this commit i see differences only in 2 benchmarks. baseline is without this
change. I am looking at the SuperChars regression.

StrToInt                              9297             8402      -895     -9.6%     1.11x
SuperChars                          676554           748375    +71821    +10.6%     0.90x (!)