IRGen: adjust for clang API change
When building with assertions, a recent change would cause the GlobalVariable
constructor to trip over the assertion:
  `InitVal->getType() == Ty`.

This was caused by the types used in swift differing from the types in clang.
In particular, the class global constructor was using `struct objc_class *` and
expecting a bitcast to coerce the type into the constructor type
(struct._class_t *).

A similar issue arose during the creation of the metaclass where the empty
vtable struct mismatched due to use of an opaque pointer, and the RO data type.

The RO type is a complicated and this change takes a shortcut of assuming that
it will be initialized by clang into the same module and querying the type
rather than trying to re-create it.