[stdlib] WIP - overlaid new indexing methodology on existing Collection
I overloaded the new indexing methodology on top of Indexable and
Collection in a passive (inactive) way. I also added basic aspects of
BidirectionalCollection and RandomAccessCollection in their own files.
Finally I applied many “TODO: swift-3-indexing-model” to things left to
work on and or review. A lot of the prototype code has not yet been
incorporated.