Remove two runs of the passes in AddSSAPasses.
Re-apply b00dcbe with a small test update, and a small change in pass
ordering.

I measure around a 10% reduction in compile times of release no-assert
builds of the stdlib and StdlibUnitTest.

For release + debug-swift builds, I see 20% reduction in stdlib compile
time.

My latest measurements show a few regressions at -O:
  Calculator
  NSError
  SetIsSubsetOf
  Sim2DArray

There is a small (0.1%) reduction in the libswiftCore.dylib size.

Being able to remove these is a consequence of the reordering that
happened in e50daa6.