[ClangImporter] Record C declarations found in @interfaces.
(including both classes and categories, and also @protocol declarations)

This is a hack in that we use two different mechanisms for importing
bridging header decls now: the manual output of ParseTopLevelDecl, and
the callbacks of a clang::ASTConsumer. We should probably unify these,
but that's a bigger change that isn't as obviously safe.

Also note that we probably aren't getting this right for things included
with -include, which we already don't have a good model for.

This adds test cases for both the module and bridging header cases,
although the former worked correctly already.

rdar://problem/24806068