Sema: Split off WitnessChecker from ConformanceChecker, NFC
With resilient protocols, we have to emit a default witness table
for SILGen, independent of any one conformance.

This patch introduces a WitnessChecker base class, which contains
parts of the ConformanceChecker responsible for lookup and matching
of witnesses.

A default witness lookup will be performed inside the protocol's
DeclContext in order to find extension methods, without opening any
associated types.