Serialize only declarations of pre-specializations.
Only declarations of whitelisted pre-specializations from with public linkage need
to be serialized as they will be used by UsePrespecializations pass during -Onone
compilations to check for availability of concrete pre-specializations.

The bodies of these functions are not required as they cannot be used anyways,
because they may refer to symbols with non-public linkage.