[stdlib] Make use of `staticReport` in `_precondition`
With this change the compiler is able to generate error messages if it is able
to verify a precodition will fail at runtime.

Only _precondition's inside @_transparent functions are able to fully leverage
this change because else the compiler can't fold the condition into a true/false
value. The only functions at the moment matching these criteria are `...` and
`..<` for which this feature is requested in SR-707.