SILGen: Simpler mangling for materializeForSet callback
We don't really need to mangle the AST type of the callback, besides
it doesn't have an AST type at all, because it is polymorphic.

Also, mangle the closure as if it were parented by the requirement
and not the witness, for consistency with witness thunks.

The mangling should still be unique since it includes the conformance.

NFC other than updating tests for new mangling.