[ReconstructType] Start sketching in lookup to get local declarations
Add cases to handle basic local variable declarations by looking up the
identifier found in the LocalDeclName. There is some duplication with
Identifier nodes that could be cleaned up still.

The actual lookup is currently dumb - it just attempts to do a
function-wide lookup at the end location of the function body. We'll
need a better way to find the local declarations, but for now it lets us
make some incremental progress.